fixed iOS video bug
when playing a video on iOS, the video is played in its own window. when the user taps 'Done' to stop the video and go back to the webpage, the play button doesn't reappear.
fix: enabling the smp cta and hiding the custom overlay on mobile devices, and disabling the smp cta and showing the custom overlay on devices wider than 748px
using '-hidden' class on overlay to hide it instead of using jquery's .hide()/.show()
rebuilt custom jquery with dimensions module included so i can use .width() when handling window resize